Added classes DAO. Fixed bugs in entity classes in fields ID.

+++ src/main/java/com/riakhin/entity/Category.java
@@ -19,7 +19,7 @@ public class Category {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "category_id")
-    private Short id;
+    private Byte id;
 
     @Column(name = "name")
     private String categoryName;
