+++ src/main/java/com/riakhin/entity/Store.java
@@ -18,7 +18,7 @@ public class Store {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "store_id")
-    private Short id;
+    private Byte id;
 
     @OneToOne
     @JoinColumn(name = "manager_staff_id")
